requirements.txt
python-telegram-bot==13.13
python-dotenv==0.20.0
redis==4.3.4